+++ src/TestTrackingDiagrams.LightBDD.Core/UnifiedReportFormatter.cs
@@ -39,7 +39,7 @@ public class UnifiedReportFormatter : IReportFormatter
         if (ExpectedTestCount != null)
         {
             var scenarioCount = features.SelectMany(f => f.GetScenarios()).Count();
-            if (scenarioCount != ExpectedTestCount())
+            if (scenarioCount > ExpectedTestCount())
                 return;
         }