+++ src/TestTrackingDiagrams.LightBDD.Core/UnifiedSpecificationsDataFormatter.cs
@@ -22,7 +22,7 @@ public class UnifiedSpecificationsDataFormatter : IReportFormatter
         if (ExpectedTestCount != null)
         {
             var scenarioCount = features.SelectMany(f => f.GetScenarios()).Count();
-            if (scenarioCount != ExpectedTestCount())
+            if (scenarioCount > ExpectedTestCount())
                 return;
         }